move RuleSnoozeBadge component to rule_management sub domain
