+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/components/rule_snooze_badge/index.ts
