+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/components/rule_snooze_badge/rule_snooze_badge.tsx
@@ -7,10 +7,10 @@
 
 import { EuiButtonIcon, EuiToolTip } from '@elastic/eui';
 import React, { useMemo } from 'react';
-import { useUserData } from '../../../detections/components/user_info';
-import { hasUserCRUDPermission } from '../../../common/utils/privileges';
-import { useKibana } from '../../../common/lib/kibana';
-import { useInvalidateFetchRulesSnoozeSettingsQuery } from '../../rule_management/api/hooks/use_fetch_rules_snooze_settings';
+import { useUserData } from '../../../../detections/components/user_info';
+import { hasUserCRUDPermission } from '../../../../common/utils/privileges';
+import { useKibana } from '../../../../common/lib/kibana';
+import { useInvalidateFetchRulesSnoozeSettingsQuery } from '../../api/hooks/use_fetch_rules_snooze_settings';
 import { useRuleSnoozeSettings } from './use_rule_snooze_settings';
 
 interface RuleSnoozeBadgeProps {
