+++ x-pack/plugins/security_solution/public/detections/components/rules/step_rule_actions/rule_snooze_section.tsx
@@ -8,7 +8,7 @@
 import React from 'react';
 import { css } from '@emotion/react';
 import { EuiFlexGroup, EuiFlexItem, EuiText, useEuiTheme } from '@elastic/eui';
-import { RuleSnoozeBadge } from '../../../../detection_engine/components/rule_snooze_badge';
+import { RuleSnoozeBadge } from '../../../../detection_engine/rule_management/components/rule_snooze_badge';
 import * as i18n from './translations';
 
 interface RuleSnoozeSectionProps {
